Fix array OOB crash when rendering sprites

+++ src/Sprites.cpp
@@ -148,10 +148,10 @@ void Sprites::drawBitmap(int16_t x, int16_t y,
             Arduboy2Base::sBuffer[ofs] = data;
           }
           if (yOffset != 0 && sRow < 7) {
-            data = Arduboy2Base::sBuffer[ofs + WIDTH];
+            data = Arduboy2Base::sBuffer[(uint16_t)(ofs + WIDTH)];
             data &= (*((unsigned char *) (&mask_data) + 1));
             data |= (*((unsigned char *) (&bitmap_data) + 1));
-            Arduboy2Base::sBuffer[ofs + WIDTH] = data;
+            Arduboy2Base::sBuffer[(uint16_t)(ofs + WIDTH)] = data;
           }
           ofs++;
           bofs++;
@@ -170,7 +170,7 @@ void Sprites::drawBitmap(int16_t x, int16_t y,
             Arduboy2Base::sBuffer[ofs] |= (uint8_t)(bitmap_data);
           }
           if (yOffset != 0 && sRow < 7) {
-            Arduboy2Base::sBuffer[ofs + WIDTH] |= (*((unsigned char *) (&bitmap_data) + 1));
+            Arduboy2Base::sBuffer[(uint16_t)(ofs + WIDTH)] |= (*((unsigned char *) (&bitmap_data) + 1));
           }
           ofs++;
           bofs++;
@@ -189,7 +189,7 @@ void Sprites::drawBitmap(int16_t x, int16_t y,
             Arduboy2Base::sBuffer[ofs]  &= ~(uint8_t)(bitmap_data);
           }
           if (yOffset != 0 && sRow < 7) {
-            Arduboy2Base::sBuffer[ofs + WIDTH] &= ~(*((unsigned char *) (&bitmap_data) + 1));
+            Arduboy2Base::sBuffer[(uint16_t)(ofs + WIDTH)] &= ~(*((unsigned char *) (&bitmap_data) + 1));
           }
           ofs++;
           bofs++;
@@ -223,10 +223,10 @@ void Sprites::drawBitmap(int16_t x, int16_t y,
             Arduboy2Base::sBuffer[ofs] = data;
           }
           if (yOffset != 0 && sRow < 7) {
-            data = Arduboy2Base::sBuffer[ofs + WIDTH];
+            data = Arduboy2Base::sBuffer[(uint16_t)(ofs + WIDTH)];
             data &= (*((unsigned char *) (&mask_data) + 1));
             data |= (*((unsigned char *) (&bitmap_data) + 1));
-            Arduboy2Base::sBuffer[ofs + WIDTH] = data;
+            Arduboy2Base::sBuffer[(uint16_t)(ofs + WIDTH)] = data;
           }
           ofs++;
           mask_ofs++;
